refactor(FeatureAcf): rename namespace
